backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.0
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
firebase-admin==6.5.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2

# Linting
ruff==0.6.9
